+++ grace-shell/src/main/groovy/org/grails/cli/ant/tasks/Gradle.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2024 the original author or authors.
+ * Copyright 2022-2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,29 +13,35 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.cli.profile.tasks;
+package org.grails.cli.ant.tasks;
 
 import org.apache.tools.ant.BuildException;
 import org.apache.tools.ant.taskdefs.ExecTask;
+import org.apache.tools.ant.taskdefs.condition.Os;
 import org.apache.tools.ant.types.Commandline;
 
 /**
  * @author Michael Yan
  * @since 2023.0
  */
-public class Run extends ExecTask {
+public class Gradle extends ExecTask {
 
     private String command;
 
-    public Run() {
+    public Gradle() {
+        if (Os.isFamily("windows")) {
+            super.setExecutable("gradlew");
+        }
+        else {
+            super.setExecutable("./gradlew");
+        }
     }
 
     public void addText(String text) {
         this.command = text;
         String[] tmp = Commandline.translateCommandline(text);
         if (tmp != null && tmp.length > 0) {
-            setExecutable(tmp[0]);
-            for (int i = 1; i < tmp.length; i++) {
+            for (int i = 0; i < tmp.length; i++) {
                 createArg().setValue(tmp[i]);
             }
         }
@@ -43,7 +49,7 @@ public class Run extends ExecTask {
 
     @Override
     public void execute() throws BuildException {
-        log(this.command);
+        log(command);
         super.execute();
     }
