+++ grace-shell/src/main/resources/org/grails/cli/ant/tasks/antlib.xml
@@ -19,12 +19,12 @@
      under the License.
 -->
 <antlib>
-    <taskdef name="file" classname="org.grails.cli.profile.tasks.File"/>
-    <taskdef name="git" classname="org.grails.cli.profile.tasks.Git"/>
-    <taskdef name="grace" classname="org.grails.cli.profile.tasks.Grace"/>
-    <taskdef name="gradle" classname="org.grails.cli.profile.tasks.Gradle"/>
-    <taskdef name="groovy" classname="org.grails.cli.profile.tasks.Groovy"/>
-    <taskdef name="npm" classname="org.grails.cli.profile.tasks.Npm"/>
-    <taskdef name="run" classname="org.grails.cli.profile.tasks.Run"/>
-    <taskdef name="say" classname="org.grails.cli.profile.tasks.Say"/>
+    <taskdef name="file" classname="org.grails.cli.ant.tasks.File"/>
+    <taskdef name="git" classname="org.grails.cli.ant.tasks.Git"/>
+    <taskdef name="grace" classname="org.grails.cli.ant.tasks.Grace"/>
+    <taskdef name="gradle" classname="org.grails.cli.ant.tasks.Gradle"/>
+    <taskdef name="groovy" classname="org.grails.cli.ant.tasks.Groovy"/>
+    <taskdef name="npm" classname="org.grails.cli.ant.tasks.Npm"/>
+    <taskdef name="run" classname="org.grails.cli.ant.tasks.Run"/>
+    <taskdef name="say" classname="org.grails.cli.ant.tasks.Say"/>
 </antlib>
